Setup visual studio to compile the extension.
Setup .gitignore to ignore a bunch of files that can be regenerated - note that the compiled binaries aren't being commited, just the code and visual studio setup to set it up (at this stage, Visual Studio 2013).

Coordinate is the project for the Functions.Geographic.Coordinate module.

Distance.cpp contains functions for determining distance.

PythonExtensionBase.props is the property sheet for a basic setup.

AL-55 #time 10m
AL-56 #time 10m