We are now able to perform a prediction based on Modifier_Base (simple distance / speed calculation).
It currently interpolates the distance between two points to fill it in with points (currently relatively low resolution), and then performs regression analysis to determine the most likely section of the path that the object is on.

We are able to output a relatively good prediction (also, Cartesian conversion inaccuracy does not seem to have to large of an effect).

AL-66 #time 3h